loaded zone textures

+++ src/ABFA/State.hs
@@ -11,22 +11,22 @@ import ABFA.Data
 -- that happens, such as window state, random seeds, screen
 -- width and height, grids and zones.
 data State = State
-  { stateGame       :: !GameState      -- the current screen
-  , stateGamePrev   :: !GameState      -- the previous screen
-  , stateSettings   :: !Settings       -- the currents settings
-  , stateWParams    :: !WorldParams    -- parameters for the world generator
-  , stateStdGens    :: ![StdGen]       -- a set of rngs
-  , stateZoom       :: !Float          -- the zoom when in the zone screen
-  , stateGrid       :: ![Int]          -- the world grid
-  , stateOG         :: ![Int]          -- the original world grid
-  , stateElev       :: ![Int]          -- the average elevation of each tile
-  , stateCursor     :: !(Int, Int)     -- the cursor position
-  , stateTime       :: !Integer        -- the time since 0
-  , stateLua        :: !Lua.State      -- the state of the lua interpreter
-  , stateShellBuff  :: ![String]       -- a string containing the shell history and prompt
-  , stateShellInput :: !String         -- the string the user is typing
-  , stateZone       :: ![Zone]         -- the zone data, packed in bytestring, the whole array at once but lazy
-  , stateZoneCam    :: (Int, Int, Int) -- the location of the camera when in the zone screen
+  { stateGame       :: !GameState          -- the current screen
+  , stateGamePrev   :: !GameState          -- the previous screen
+  , stateSettings   :: !Settings           -- the currents settings
+  , stateWParams    :: !WorldParams        -- parameters for the world generator
+  , stateStdGens    :: ![StdGen]           -- a set of rngs
+  , stateZoom       :: !Float              -- the zoom when in the zone screen
+  , stateGrid       :: ![Int]              -- the world grid
+  , stateOG         :: ![Int]              -- the original world grid
+  , stateElev       :: ![Int]              -- the average elevation of each tile
+  , stateCursor     :: !(Int, Int)         -- the cursor position
+  , stateTime       :: !Integer            -- the time since 0
+  , stateLua        :: !Lua.State          -- the state of the lua interpreter
+  , stateShellBuff  :: ![String]           -- a string containing the shell history and prompt
+  , stateShellInput :: !String             -- the string the user is typing
+  , stateZone       :: ![Zone]             -- the zone data, packed in bytestring, the whole array at once but lazy
+  , stateZoneCam    :: (Float, Float, Int) -- the location of the camera when in the zone screen
   }
 
 -- the gamestate controls which screen we are currently drawing
